update gear shift subsystem
